Update new music grid layout parameters

+++ lib/widgets/pages/browse_sections/new_music_grid.dart
@@ -42,7 +42,7 @@ class _NewMusicSectionState extends State<_NewMusicSection> {
     // Keep the carousel compact: at most three columns, while each column
     // retains the existing four-row layout.
     final pageCount =
-        ((widget.songs.length / _kItemsPerPage).ceil()).clamp(1, 3) as int;
+        ((widget.songs.length / _kItemsPerPage).ceil()).clamp(1, 3);
 
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
